refactor: 프로필 Contribution 통계 영역 컴포넌트로 분리
